// frontend/src/api/generated/models/batch-execution-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BatchExecutionStatus } from './batch-execution-status';
// May contain unused imports in some cases
// @ts-ignore
import type { BatchStepStatus } from './batch-step-status';

/**
 * 
 * @export
 * @interface BatchExecutionDetail
 */
export interface BatchExecutionDetail {
    /**
     * バッチ実行ログID
     * @type {number}
     * @memberof BatchExecutionDetail
     */
    'id'?: number;
    /**
     * 処理対象営業日
     * @type {string}
     * @memberof BatchExecutionDetail
     */
    'targetBusinessDate'?: string;
    /**
     * 
     * @type {BatchExecutionStatus}
     * @memberof BatchExecutionDetail
     */
    'status'?: BatchExecutionStatus;
    /**
     * 
     * @type {BatchStepStatus}
     * @memberof BatchExecutionDetail
     */
    'step1Status'?: BatchStepStatus;
    /**
     * 
     * @type {BatchStepStatus}
     * @memberof BatchExecutionDetail
     */
    'step2Status'?: BatchStepStatus;
    /**
     * 
     * @type {BatchStepStatus}
     * @memberof BatchExecutionDetail
     */
    'step3Status'?: BatchStepStatus;
    /**
     * 
     * @type {BatchStepStatus}
     * @memberof BatchExecutionDetail
     */
    'step4Status'?: BatchStepStatus;
    /**
     * 
     * @type {BatchStepStatus}
     * @memberof BatchExecutionDetail
     */
    'step5Status'?: BatchStepStatus;
    /**
     * 
     * @type {BatchStepStatus}
     * @memberof BatchExecutionDetail
     */
    'step6Status'?: BatchStepStatus;
    /**
     * エラーメッセージ（FAILED時のみ）
     * @type {string}
     * @memberof BatchExecutionDetail
     */
    'errorMessage'?: string | null;
    /**
     * 処理開始日時
     * @type {string}
     * @memberof BatchExecutionDetail
     */
    'startedAt'?: string;
    /**
     * 処理完了日時（RUNNING中はnull）
     * @type {string}
     * @memberof BatchExecutionDetail
     */
    'completedAt'?: string | null;
    /**
     * 実行者氏名
     * @type {string}
     * @memberof BatchExecutionDetail
     */
    'executedByName'?: string;
}