// frontend/src/api/generated/models/batch-execution-page-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { BatchExecutionDetail } from './batch-execution-detail';

/**
 * 
 * @export
 * @interface BatchExecutionPageResponse
 */
export interface BatchExecutionPageResponse {
    /**
     * 
     * @type {Array<BatchExecutionDetail>}
     * @memberof BatchExecutionPageResponse
     */
    'content'?: Array<BatchExecutionDetail>;
    /**
     * 
     * @type {number}
     * @memberof BatchExecutionPageResponse
     */
    'page'?: number;
    /**
     * 
     * @type {number}
     * @memberof BatchExecutionPageResponse
     */
    'size'?: number;
    /**
     * 
     * @type {number}
     * @memberof BatchExecutionPageResponse
     */
    'totalElements'?: number;
    /**
     * 
     * @type {number}
     * @memberof BatchExecutionPageResponse
     */
    'totalPages'?: number;
}