// frontend/src/api/generated/models/batch-execution-status.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * バッチ実行ステータス - SUCCESS: 正常完了 - FAILED: 異常終了 - RUNNING: 実行中 
 * @export
 * @enum {string}
 */

export const BatchExecutionStatus = {
    Success: 'SUCCESS',
    Failed: 'FAILED',
    Running: 'RUNNING'
} as const;

export type BatchExecutionStatus = typeof BatchExecutionStatus[keyof typeof BatchExecutionStatus];